Do not delete the kube-apiserver explicitly
While we delete all services and volumes and also check whether resources still exist, the kube-controller-manager may still be processing the deletion in the background. The fact that no more services or PVCs/PVs are known to the kube-apiserver does not mean that the underlying infrastructure resources are already gone. Thus, we keep the kube-apiserver deployment to keep the kube-controller-manager functional and allow a proper cleanup.