Refresh the `.status.operationStartTime` on new operations
The time is only refreshed if the last operation was succeeded, the specification changed or no previous timestamp given.
When the last operation ended with state "Error" it won't be refreshed so that one can determine the time when the operation should not longer be retried (maximum duration is configurable in `ControllerManagerConfiguration`).