DNS section allows referencing a secret containing the credentials
In case no default domain has been specified, the standard way of finding the credentials for the DNS provider is to look up the cloud provider secret referenced in a Shoot for the required credentials.
However, when a user has its own domain and wants to use it multiple times in its namespace/project, it is more convenient and maintainable to only declare the DNS credentials secret once. With this commit it is allowed to optionally specify the .spec.dns.secretName field (only when .spec.dns.hostedZoneID and .spec.dns.domain are specified as well). The Gardener will then skip looking up the cloud provider secret but only make usage of the secret referenced in .spec.dns.secretName.