Gardner Vagrant Provider implementation and docs
The `gardener-vagrant-provider` executable is a simple `GRPC` service, which
runs on the developer's machine and it's responsible for creation of Shoot
cluster's VMs. It's main target is to test the Gardener without the need to get
a Cloud Provider account and credentials.

The `gardener-controller-manager` now can talk to this service and provision
/ delete machines during the `DeployInfrastructure` / `DestroyInfrastructure` step.

The development setup for `gardener-vagrant-provider` adds an additional optional
step to the local dev setup.