ShootValidator admission plugin only allows changing values when they are part of referenced CloudProfile
As @mandelsoft suggested (and that's how it's implemented right now; similar to what we (@vlerenc) discussed about yesterday): Instead of preventing the CloudProfile from removing values in the constraints which are still in use by any Shoot, we should rather adapt the existing ShootValidator admission plugin. With this commit the folllowing behaviour is implemented: When a field X in a Shoot manifest is changed to value Y,  the operation will only succeed if Y is part of the referenced CloudProfile. For existing Shoots whose field X has another value Z that is not part of the CloudProfile, updating other fields different than X does not prevent the request from succeeding (like it was before). This allows us to remove old entries from the CloudProfiles (so that no new Shoots can use old values).

closes #82