Successful Shoot deletions wait until resource changed have been applied
We have seen cases where the updated Shoot status or deleted finalizer were not reflected in the cached version of the Shoot resource (although the API calls succeeded). The next worker picking the Shoot key out of the queue tried to delete the Shoot again, starting with updating the status last operation to "delete processing". In the meanwhile, because of the fact that the finalizer has been deleted, the Shoot was garbage collected. Correspondingly, the currently running worker failed updating the Shoot resource, resulting in unwanted errors.