Kubernetes API discovery is retried with an exponential backoff
When bootstrapping a new Shoot cluster, the Shoot's Kubernetes API server may take some time to create all the APIService objects for its API groups. We need to have some retry logic to overcome that delay. Unfortunately, there is no clear indication when the API server is really ready to serve requests as one is used to see it when it runs already for a period of time.