Include Heapster specification hash into its deployment name
We manage the addons in the Shoot via the kube-addon-manager and usually we do that by labelling the respective resources with the "Reconcile" action. However, for the Heapster deployment we couldn't use "Reconcile" but only "EnsureExists". The reason for that is the heapster-nanny sidecar container in that deployment/pod which dynamically adjusts the resource requests of the deployment. In the end, the kube-addon-manager would in every resync period discard the resource requests set by heapster-nanny (because those are not reflected in the kube-addon-manager's desired state of that deployment); resulting in an alternating update cycle of Heapster's deployment manifest.

By including a hash value of the deployment specification we can make sure that updates/changes made by us in future Gardener versions are correctly brought to the existing Shoot clusters.