Kubelet TLS bootstrapping and certificate rotation
More infos: https://kubernetes.io/docs/admin/kubelet-tls-bootstrapping/

Notes: Ideally, every node should have its own single bootstrap token which is only valid for a short period of time. However, to implement this we need to strongly align with machine-controller-manager implementation and move some parts of the cloud config there (?). This commit is considered as first step to have TLS bootstrapping and certificate rotation in place, but it should be improved further.