Do not longer run kubelet inside a container
We have experienced a lot of problems where the kubelet was not able to unmount volumes mounted into the pods. We decided that it's better for our scenario to just start the kubelet as normal process on the hosts. Running it inside a container doesn't bring us much benefits either. Also, the `socat` binary is now installed on the hosts.

cc @mvladev