Validate Shoot/project name length sum
We use the identifier "shoot-<project-name>-<shoot-name> in nearly all places: when creating infrastructure
resources, Kubernetes resources, DNS names, etc. Some infrastructure resources have length constraints that
this identifier must not exceed 30 characters, thus we need to check whether Shoots do not exceed this limit.
The project name is a label on the namespace. If it is not found, the namespace name itself is used as project
name.