Deploying AWS Shoots into existing VPCs requires attached internet gateway
Before this commit, when deploying an AWS Shoot cluster into an existing VPC, the Gardener was checking whether the given VPC had an internet gateway attached. If no, then one was created. However, with the introduction of the reconcile logic, the internet gateway would be removed in the next reconciliation phase (because now there is one; following the relevant Terraform resource would not have been written into the computed Terraform script). This was leading to an alternating create/delete procedure.
Now the Gardener requires that the user has attached an internet gateway to the existing VPC. It will only create one when it also creates the VPC itself.

cc @mandelsoft @frnksgr @afritzler