Gardener labels Shoots which are unhealthy
In order to allow easy identification of Shoot clusters with issues the Gardener will now label those Shoots with `shoot.garden.sapcloud.io/unhealthy=true`. If the cluster is healthy the label will not be present.
The health check is done by the Shoot care controller which skips Shoots having an ongoing operation/reconciliation process. The normal Shoot controller writes the same label if its last executed operation was not successful.

A Shoot cluster is considered unhealthy if one of the following conditions match:
- The `.status.lastOperation.state` field not equals `Succeeded`.
- The `.status.lastError` field is not empty.
- The `ControlPlaneHealthy` condition status not equals `True`.
- The `EveryNodeReady` condition status not equals `True`.
- The `SystemComponentsHealthy` condition status not equals `True`.

cc @grolu @petersutter @DockToFuture