Lower default retry times in creation/reconciliation/deletion flow
The long default retry times where introduced when we had not reconcilitation/update mechanism at all. Now that we have it since a long time we can lower the times.
When a flow fails, the Shoot is added to the queue after 15s anyway and the Gardener will restart the flow again.