Kubelet monitoring script reboots the VM if ready state toggles too often
The monitoring script installed on the VMs has checked the `/healthz` endpoint of the running kubelet so far. Besides of doing that, it now also checks the `Ready` condition of the corresponding Node object. If its status toggles too often between True and False (currently 5 toggles in 10 minutes) the VM will be rebooted.