Better handling of retry cycle start time
The Gardener controller-manager can be configured to retry a create/reconcile/delete operation for a certain time. If the operation was not successful within this time frame, it sets the Shoot status to "Failed". After that, only specification changes or re-triggering the Shoot deletion force the controller-manager to start over a new operation.

- Renamed .status.operationStartTime to .status.retryCycleStartTime