Allow to retry failed operations
So far, when a Shoot cluster state has been set to "Failed", only changing the specification or putting the confirmation annotation again would trigger the Gardener to startover a new operation. With this commit we have introduced a new annotation called "shoot.garden.sapcloud.io/operation". When it is set with a value equals "retry" for a Shoot cluster in a Failed state then the Gardener API server will increment the .metadata.generation.
The controller-manager will skip Shoots in a Failed state whose .metadata.generation equals .status.observedGeneration. Thus, after the API server has incremented it, the controller-manager will startover a new operation without the necessity of changing the spec or doing something else.

cc @mandelsoft