Delete kube-apiserver when destroying the internal DNS record on Shoot deletion
Previously, when the Gardener has started destroying the internal DNS record (kicking off the Terraform job) and crashed before it could see that it was completed, it could not finish the Shoot deletion after restart. The reason for that was that the internal DNS record was gone, however, the namespace in the Seed cluster was not deleted and therefore the kube-apiserver deployment was still running. Consequently, the Gardener thought that it needs to clean the resources, however, due to the meanwhile missing DNS record, it could not construct a client to talk to the kube-apiserver. The deletion was stuck at this point.
Now, when we have deleted all the infrastructure resources, it's safe to delete also the kube-apiserver deployment (while we kick off the job destroying the internal DNS record in parallel). In case it crashes/looses connection/need to startover for other reasons, it won't try to cleanup the Kubernetes resources and go ahead with deletion the namespace.

cc @gonzolino