Create Prometheus for seed kubelet scraping
Added a Prometheus for the kube-system of a Seed cluster. It collects metrics
from kubelets (+cadvisor) to reduce the load on the Prometheus shoot instances.
The network traffic is reduced, because the filtering is done when the Shoot
Prometheus uses the /federation endpoint with filter options.

To be 100% sure that Promethus can talk to the Shoot kubelets, the API server
proxy functionality is used.