Faster way to shut down controller manager
So far, we have waited for all controllers to terminate, however, that implementation arose during a time when the Gardener was not able to reconcile or update Shoots. Nowadays, the Gardener is able to terminate at each point in time and can start and reconcile Shoots independent of the state they were left at.

closes #130