Explicitly delete AWS ELBs and security groups created from Kubernetes
This is a mitigation for kubernetes/kubernetes#17626 and the not yet merged pull requests kubernetes/kubernetes#54569 and kubernetes/kubernetes#65912.
The kube-controller-manager does not add a finalizer to `Service` objects and may forget that it has created resources in the infrastructure due to this.

closes #129

```improvement user
Gardener does now explicitly delete the load balancers and security groups belonging to Kubernetes services for AWS Shoots. This is to mitigate the issue that the kube-controller-manager does not use finalizers on Service objects and may forget that it has created resources in the underlying infrastructure.
```