Logging stack for shoot control plane.
The logs for the seed are gathered in a separate Elasticsearch instance in the
garden namespace.  The logs for each shoot are gathered in its own Elasticsearch
instance in the shoot namespace.

The logs are initially gathered by a daemon set of fluent-bit in the seed
cluster. Then they are forwarded to a single (central) instance of fluentd. This
central fluentd then forwards each log message to its Elasticsearch instance.

Application logs are of no concern.