Add DeletionConfirmation admission controller
The current way of deletion of Shoot resources is a left over from when the
Gardener was working as TPR operator and external validation webhooks were not
available.

DeletionConfirmation admission controller fixes this issue by allowing deletion
of resources to happen when a Shoot resource is annotated with
`confirmation.garden.sapcloud.io/deletion=true`

It is disabled by default to preserve the old behavior.