Only patch the checksum annotations instead of redeploying in Shoot deletion
See also: e671c15d26aa76e7c698355272867d940f6b95b3, 77023be55eb1cfb85d4b11904afd25535d6009c2

We can't redeploy the cloud provider config as the Terraform state may
not longer contain the needed output variables. Hence, we need to
refresh the cloud provider credentials (that may be part of the cloud
provider config for some clouds) instead of recomputing it from scratch
(which might not be possible for the said reason). Moreover, it's safer to just
update the checksum of the cloud provider secret in the annotations of
the API server and the controller manager instead of redeploying both
(as their configuration might depend on other operations that are only
part of the reconciliation and not of the deletion flow).