Rely on Kubernetes garbage collection to cleanup resources with owner references
When deleting a cluster Gardener cleans up all workload resources and all of those which may could have
caused a footprint in the underlying infrastructure. It maintains a list of exceptions to keep the cluster functional
in order to allow controllers/pods to terminate gracefully (e.g., calico or kube-dns are not getting deleted). However,
Gardener only knows the name of the kube-dns deployment or the calico-node daemonset; it does not know the name
of the corresponding created pods. As the delete-and-wait logic for resource clean up is generic and as we can't check
efficiently whether the "root" resource of a chain of resources with owner references we do now skip those which have
been created by others. With that we rely on the controllers which created them, and on Kubernetes garbage collection
to clean up properly.