Do not set --apiserver-count on kube-apiserver
The kube-apiserver's advertise address is the IP address of the load balancer backing it. If we have multiple API server pods, all of them will still advertise the same load balancer IP address, so we do not get in trouble with the master reconciler if we simply don't specify the flag (defaults to 1).

/cc @ggaurav10