Remove support of watching only one specific namespace instead of the whole cluster
This "feature" did exist for historic reasons (when the Gardener was
just a CRD-operator but no full-blown API server extension with
a controller manager), and it was anyway only implemented for Shoot
resources but not for all the others. It doesn't make sense to maintain
this dead code as it's anyway not very useful anymore (as you can only
have one Gardener per Kubernetes cluster).