White list of all Prometheus metrics.
When scraping individual targets Prometheus stores only a collection of
white-listed metrics. This reduces the samples stored and improves performance
and memory utilization.

Endpoints using the `prometheus.io/*` annotations are now limited to `500`
samples.

Refactoring also allowed for `cAdvisor` networking metrics to be scraped for
Pods and `systemd` containers are again shown in the Grafana Dashboard.

Config reloaders for Prometheus and blackbox-exporter are dropped from the
Shoot's Prometheus.