Fix CoreDNS PTR Records resolution
Currently with CoreDNS it is possible to resolve A records for services
however it was not possible to to do reverse lookups.

This commits fix the issue that prevented reverse-lookup configuration
to be applied (basically resolving the config between variables
overriden from the golang code injection vs. direct rendering from
values).

```noteworthy user
A bug in the CoreDNS configuration that prevented reverse DNS lookups has been resolved.
```