Increase kubelet's API/event recording QPS/burst values
The client-go library which is used by kubelet to perform requests
against the kube-apiserver throttles requests if they go above the
configured thresholds. For kubelet, the default qps=burst=5, so it's
very low. If many pods get scheduled at the same time then the kubelet
cannot update the pods' statuses because it got throttled by the client
when trying to do so. However, it did correctly start the pods.

```improvement user
The kubelet's client QPS/burst values have been increased in order to allow the kubelet to faster react on many pod creations at the same time.
```