Add InternalIP cache to Kubelet Monitor Service
Let Kubelet Monitor Serivce use a cache for the InternalIP of the
Host / Node. This can be used to update the K8s Node object in case
the Kubelet Service stopped exposing it. Only restart Kubelet Service
if cache is out-dated or empty.

```improvement user
An IP cache was added to the kubelet monitoring script deployed to each shoot worker node. It should prevent unnecessary restarts of the kubelet in case it loses the VM's internal IP address.
```