Fix dashboard's label selectors
In `apps.v1` Deployment's `spec.labelSelector` is immutable.
Picking correct labels for selection should be done with care as we
might not be able to change them in the future.

If versions are needed in `Deployments` then `matchExpressions` with
`operator` `Exists` should be used:

```yaml
spec:
  selector:
    matchExpressions:
    - {key: app, operator: In, values: [nginx] }
    - {key: version, operator: Exists }
```

As a temporary fix the `apiVersion` for the Dashboard's Deployment is
now set to `extensions/v1beta1` where it can be modified.

/cc @gardener/gardener-maintainers