Extend Nginx-Ingress addon by DNS settings
Extend Nginx-Ingress addon by DNS settings which specify the ingress
domains for a Shoot cluster.

`standardRecords` are supposed to be chosen by Shoot owners and must be
part of the Shoot domain. This version allows to set a single domain
only.

`additionalRecords` are set by Gardner only. For a first version, a
single domain is randomly picked but suffixed by the internal domain
of Gardener.