Add spec hash labels to ControllerInstallations
Similar to what the kube-controller-manager does when creating a `Pod` resource (injecting a hash of the pod-template into the `Pod`s labels: `pod-template-hash=<hash-value>`) we do now inject a hash for the specifications of the referenced objects (seed, registration).
This allows others to easily identify whether something relevant has changed within the referenced objects.