VM user-data does no longer contain the entire cloud-config-downloader
Most cloud providers have a limit of 16 KB regarding the user data that my be sent during VM creation.
Recently, we have split the kubeconfig into individual parts (ca.crt, server, cllient.crt, client.key) to get rid of kubectl in the cloud-config-downloader script.
The entire kubeconfig was only used later to extract the server name and the ca.crt - this is not needed any more.

With this change the user data has been shrinked to ~8 KB which is no longer risking that we hit the cloud providers limits.