Remove usage of initializers and adapt PV-labelling logic
Due to kubernetes/kubernetes#72972 the alpha `Initializers` feature will be entirely removed with Kubernetes 1.14. We are currently relying on it as the cloud-controller-manager uses it to label persistent volumes before they become visible in the API. This functionality is also removed (see kubernetes/kubernetes#73102) with 1.14. As a consequence, in order to use PV labelling until the community has migrated all CCMs to use `MutatingWebhookConfiguration`s for that, the only way is to enable the (already deprecated) `PersistentVolumeLabel` admission controller in the kube-apiserver. That means that the kube-apiserver needs to talk to the cloud provider to figure out the region/zone of the PV (and later add them as label).
--> We have to provide the cloud credentials + config to the API server in this case.

Shoot clusters with Kubernetes version < 1.14 still have an enabled initializer feature.