Vendor gardener-extensions repository to get provider specific Go types
We are still operating on garden.sapcloud.io/v1beta1.Shoot, so we have
to migrate to the needed `InfrastructureConfig` (and in the future also
`Infrastructurestatus`) ourselves until we have changed to
core.gardener.cloud/v1alpha1.Shoot. For this we need the Go types as
well as some helper functions which we can re-use.
This vendoring is only temporarily until we have switched to the new
Shoot resource.