Add provider info to shoot namespaces in seed
With extensibility the external provider controllers need to register
certain webhooks, e.g. to add their provider-specific flags to the
kube-controller-manager or others. In order to know for which of the
many shoot namespaces in the seed they need to act Gardener does now
provide this information as labels on the namespaces. With that, the
webhooks can be easily configured using label selectors.

cc @stoyanr