Terraform creates role instead of relying on ClusterRole
Previously, Gardener created a ClusterRole for the Terraformer (once in
the seed bootstrapping process) that was bound via a RoleBinding to
a particular Terraformer in the shoot namespace. This makes it
hard/cumbersome to reuse the `terraformer` package (e.g., for
extensions).

With this change Gardener does create a Role in the namespace before
starting the Terraformer.