Allow multiple machine images
```noteworthy operator
When multiple machine images are specified in the `CloudProfile` and a `Shoot` is created without specifying a machine image then the first one in the `CloudProfile` is always chosen and considered to be the default. Previous versions of Gardener prohibited creating `Shoot` resources without specifying a machine image when the referenced `CloudProfile` did contain more than one machine image.
```