Remove kube2iam methods from CloudBotanist interface
kube2iam is deprecated and we will keep it for a short while for backwards compatibility reasons. It never fitted the `CloudBotanist` interface. Our goal with extensibility is to get rid of the `CloudBotanist` interface, hence, we now remove these kube2iam methods from it and keep them as regular functions in the `awsbotanist` package until we finally entirely remove it from Gardener.