Allow to configure backup profile in shoot-as-seed annotation
```improvement operator
It is now possible to configure the backup profile via the `shoot.garden.sapcloud.io/use-as-seed` annotation. Please set `backup.provider`, `backup.region`, `backup.secretRef.name`, `backup.SecretRef.namespace` if you want to configure it. If no such configuration is found the default behaviour is that the same provider will be used for backup. In order to explicitly disable backup for a seed configure `backup.provider=none` in the annotation.
```
```action operator
If you use have a shooted seed of provider OpenStack then you have to make sure that the referenced provider secret contains a key named `authURL` with the keystone URL. This is due to https://github.com/gardener/gardener/pull/1128.
```