Store audit policy config map resource version
Previously, the shoot configmap controller was enqueuing all referred
configmaps during start-up. In its reconciliation loop it annotated the
shoot with the operation annotation. If reconciliation was only enabled
in the maintenance time window this led to undesired reconciliations
only because GCM was (re)started.
Now, we store the resource version of the config map (similar to how we
store the resource version of the controller registration in the
controller installation). The reconciliation loop will update the seen
resource version which will only trigger a shoot reconciliation if the
configmap changed. Note that also added labels/annotations to the
configmap would trigger a reconciliation.

Co-Authored-By: Tim Usner <tim.usner@sap.com>