Use /readyz endpoint for 1.16 kube-apiservers
Also, we set the `--livez-grace-period` to `1m` which is the
maximum amount of time it should take for the kube-apiserver to complete
its startup sequence and become healthy. From kube-apiserver's start
time to when this amount has elapsed, `/healthz` will assume that
unfinished post-start hooks will complete successfully and therefore
return true.
However, `/readyz` will only return true if the post-start hooks are
finished and the kube-apiserver is ready to receive requests. As
`/readyz` also includes a shutdown check the kube-apiserver will try to
process existing requests (for the duration of request timeout) before
it kills the kube-apiserver process.

See kubernetes/kubernetes#78458