Added dependency-watchdog as a bootstrap component
Both for watching etcd and kube-apiserver endpoints
as well as for probing the shoot kube-apiserver for
loadbalancer issues.

Enhanced seed reconcilation to delete existing
dependency-watchdog deployments in the
control-plane namespaces. This code can be removed
in a future release.

Bumped dependency-watchdog to 0.4.0.

Increased node monitor grace period to give the
dependency watchdog probe a change to scale down
the kube-controller-manager before the nodes are
marked as `NotReady`.

Added a networkpolicy allow-to-all-shoot-apiservers to
global-network-policies.