Keep using hyperkube for binaries on worker nodes
We cannot just download binaries from `storage.googleapis.com` as this
would break private cloud and on-prem deployments where this location is
not accessible. Also, for these scenarios, we don't have a place where
to put the binaries on the worker nodes, and we probably don't want to
introduce this new dependency. One suitable solution would be to have
a container image like `hyperkube` that contains all the necessary
binaries, and during node-bootstrapping these binaries can be copied
over to the file system of the VM. As the Gardener project does not yet
have such a container we keep using `hyperkube` for now until we can
replace it with one built by us.