ChatApplier now takes interface{} as value
Reworked various parts of the public interface of ChartApplier to make
it more composable. Extra methods `With...` are now removed and
customizaiton can be applied with implementing `ApplyOption` and
`DeleteOption` iterterfaces.

Both `Delete` and `Apply` now can use `interface{}` instead of
`map[string]interface{}` as values. With this change is now possible to
pass a predefined struct such as

```go
type Foo struct {
        Baz string `json:"baz,omitempty"`
        Foo bool `json:"foo,omitempty"`
}
```

to it.