:bug: Fix the restoration failure alert for etcd
cc @shreyas-s-rao @wyb1 
Count is passive entity. Where as rate is active entity. Rate actually notiifies the events that count has been changed. With existinig rule, even if once the restoration is failed, the alert will be raised for lifetime until the count is reset which will occure only when container restarts.