Add managed istio feature gates for Seed clusters.
A customized installation of Istio can now be enabled and installed by
gardenlet. The feature gate is `ManagedIstio` and can be configured in
the component config.

Disabling this feature once installed does not removes any installed
resources as it might accidentally remove already existing installation
of Istio.

The installation deploys istiod in the `istio-system` namespace and
istio ingress gateway in `istio-ingress` namespace. mTLS is enforced and
`Services`, `VirtualServices` and `DestinationRules` are not exported
and advertised by default and therefore it must be explicitly enabled
either via the `networking.istio.io/exportTo: "*"` annotation or
`.exportTo: ["*"]`.

Telemetry is disabled.