Add event controller
The goal of the event controller is to persist Shoot-related events
longer for debugging reasons.

The controller deletes all non-Shoot events after the configured
`ttlNonShootEvents`.
Shoot-related events are not touched by the controller and only eventually get
deleted after the configured `ttl` in the kube-api-server configuration.

Considered alternatives to this controller were:
1. export relevant events from etcd to another persistent store.
2. extend the global `ttl` in the api-server for all events

While option 1. would be preferable it imposes dificulties as with the Control
Plane Migration the shoot CP(with the events) can move between Seed worker
nodes.

In addition this commit:
- Replaces some Shoot-related events with logs to reduce the load written to etcd
- Creates more fine grained events regarding shoot cluster maintenance and hibernation