Delete VolumeAttachments on hibernation
```improvement operator
gardenlet is now deleting all VolumeAttachments on hibernation. As during hibernation machine-controller-manager performs a "force" deletion of machines and does not wait for volumes to detach, kube-controller-manager is not able to delete the corresponding VolumeAttachments (and also the external-attacher to notice this deletion and remove its finalizer from the VolumeAttachment). Deleting VolumeAttachments on hibernation should prevent VolumeAttachments to be orphaned. Currently in the upstream kube-controller-manager, there is no garbage collection for VolumeAttachments (see [kubernetes/kubernetes#77324](https://github.com/kubernetes/kubernetes/issues/77324)).
```

Signed-off-by: ialidzhikov <i.alidjikov@gmail.com>