Fix coredns network policies.
Now they allow traffic from the Pod network. This is needed because
traffic might come from the calico tunnel IP which is part of the Pod
network.

Before this, Pods with `hostNetwork: true` and `dnsPolicy: ClusterFirstWithHostNet`
could not talk to coredns if they are running on a different Node than
CoreDNS.