Refactor gardener-resource-manager lifecycle
Before this commit the grm was deployed via helm charts. There
were two separate folders with charts for the two uses of the
grm (shoot controlplane and seed).
With this commit a go client is used to create the resources.
Also tests are added and there exists only only one interface
for both uses of the grm.