Adjust minimum resource limits for ingress Controller
We have observed on large landscapes (approx. 200 shoots). That the nginx
requires 2 GB memory. Therefore, the current limits led to OOMKills of the
Ingress Controller on startup. With all pods being unavailable and a
poddisruptionbudget of 1 the vpa was not able to delete a pod and
apply his higher recommendations.

With this commit we increased the initial requests and limits so that even
on large landscapes the ingress Controller can start. We also decreased the
minAllowed values so that on smaller landscapes the vpa is able to downscale
according to load.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>