konnectivity-server is separate deployment when SNI is enabled
Shoot's kube-apiserver scales automatically with load. On scale-up
events it could lead a situation where existing konnectivity-agents are
not connected to the konnectivity-server sidecar for that API server.

This happens for some period of time until other konnectivity-servers are restarted
by replica-reloader.

By moving konnectivity-server in a dedicated Deployment,
konnectivity-agents connections are not closed when scale up/down of
the kube-apiserver happens.