Make ingress domain for Seeds immutable
Changing the ingress domain on running Seed clusters can cause
a temporary unavailability of the Shoots monitoring.

We decided to not support the use case of changing the
ingress Domain for Seeds. This is in line with the
Shoot resource where the ingress domain is also immutable.

Add immutable description to domain fields

Cherry-pick https://github.com/gardener/gardener/pull/3394

```noteworthy operator
The ingress domain configuration for Seeds is now immutable.
```