Fix error during creation of fresh seed clusters
With the introduction of the nginx ingress controller an instance
of the the dnsProvider crd is created (or deleted) during
Seed bootstrap.
This commit fixes a bug where the instance of the crd is created
before the actual crd was registered. This leads to an error during
the Seed bootstrap.