Fix seed ingressdomain validation
Fixes a bug that prevents the Seed ingressDomain from being migrated
between dns.IngressDomain and ingress.Domain.
Currently, the validation will always throw an error. With this fix
the validation only throws an error if the value of the ingressDomain
actually changes.

```bugfix operator
An issue in the API validation has been fixed which prevented the managed ingress feature for seeds being enabled.
```