Use `patch` to update Seed.Status
Before the seed status was updated using optimistic locking which
could result in ongoing retries if different clients were updating
the Seed object.
As an optimization we are now switching to a `patch` call which
will send only the diff to the apiserver. We ensure that other conditions
are not changed by requesting the strategy `StrategicMergePatchType`.

Co-authored-by: Tim Ebert <tim.ebert@sap.com>