Restricted reversed vpn to only kubernetes 1.18 and higher.
Unfortunately, the api aggregation in kubernetes 1.16/1.17 does not respect the egress configuration of the kube-apiserver. Therefore, kube-apiserver calls to aggregated apis will fail.
In our scenario, the metrics-server runs in the  shoot cluster data plane. Its readiness/health check require direct connectivity from the kube-apiserver to the metrics-server in kubernetes 1.16/1.17.
This is why this change disables the reversed vpn for kubernetes 1.16/1.17.