Enh/shoot state deletion (#4100)
* Add ShootState to deletion confirmation

* delete ShootStateDeletionValidator

* Delete ShootState explicitly in shoot deletion flow

Before this change the Shootstate was deleted via owner references.
This led to problems when deleting and creating a Shoot with the same name
in quick succession as our admission plugin disallowed the deletion of the
ShootState.

Now the admission plugin is removed and the ShootState is deleted
explicitly as part of the shoot deletion flow. To protect from unintentional
deletion a deletionconfirmation is needed to delete the ShootState.