Switch from MergePatchOrCreate to GetAndCreateOrMergePatch
For the reasons explained in #4057, MergePatchOrCreate is not safe to use
without reading objects from the API server / cache first. Ideally we would
switch to server-side apply to simplify such operations on controlled objects
(e.g. extension resources), though we cannot leverage it for now because of
several issues described in #4122.