Remove *PatchOrCreate funcs
As described in #4057, *PatchOrCreate is easy to use incorrectly.
Remove *PatchOrCreate funcs in order to minimize the risk of misuse.
There might still be good use cases for this mechanism, though they
are limited. If there is such a good use case, we can use it also
without a helper func.