Move timestamp annotation check to WaitUntilObjectReadyWithHealthFunction
This prefactor moves the timestamp annotation check (which was introduced
as a stale check read detection mechanism) to the more generic
WaitUntilObjectReadyWithHealthFunction helper func. This allows us
to reuse the related code for components that wait for objects
that don't implement the extensionsv1alpha1.Object interface (e.g.
DNS objects).