Return faster when waiting for DNS{Provider,Entry}
In #4161 the behavior of the DNS components was changed in the case
a dns resource is in Error on Invalid state (errors are not surfaced
as fast as before, if no error code is detected).
This commit brings back the old behavior and thereby brings the dns
components even closer to the behavior of all other extension
components.