Refactor special handling for ErrorWithCodes
This prefactoring cleans up the special handling for ErrorWithCodes
in WaitUntilObjectReadyWithHealthFunction.
Before, the wait func returned MinorOrSevereError if the health
func returned an instance of ErrorWithCodes, e.g. in order to faster
return from waiting for extension resources with user errors.

This makes the handling more generic by introducing the concept
of retriable errors and using MinorOrSevereError for any error
marked as retriable. By this, health funcs that check resources
for which we don't want to detect error codes (e.g. Etcd), can
simply use retry.RetriableError to indicate that the given error
should not be treated as severe immediately without returning an
ErrorWithCodes.