Cleanup ErrorWithCodes handling in CheckDNSObject
This commit cleans up the dirty fiddling with ErrorWithCodes in the
DNS components and rather uses the cleaner approach with retriable
errors introduced in the previous commit.
As retrying a wait operation in WaitUntilObjectReadyWithHealthFunction
is not bound to returning an ErrorWithCodes anymore but rather to the
retriableError type, we can use that to indicate more explicitly, that
we want to retry waiting for a DNS object if it is in Error or Invalid
state.
This is now easier to understand as its cleaner and more explicit.