Make etcd component ready for cached client (#4197)
* Make Etcd component ready for cached client

* Refactor special handling for ErrorWithCodes

This prefactoring cleans up the special handling for ErrorWithCodes
in WaitUntilObjectReadyWithHealthFunction.
Before, the wait func returned MinorOrSevereError if the health
func returned an instance of ErrorWithCodes, e.g. in order to faster
return from waiting for extension resources with user errors.

This makes the handling more generic by introducing the concept
of retriable errors and using MinorOrSevereError for any error
marked as retriable. By this, health funcs that check resources
for which we don't want to detect error codes (e.g. Etcd), can
simply use retry.RetriableError to indicate that the given error
should not be treated as severe immediately without returning an
ErrorWithCodes.

* Cleanup ErrorWithCodes handling in CheckDNSObject

This commit cleans up the dirty fiddling with ErrorWithCodes in the
DNS components and rather uses the cleaner approach with retriable
errors introduced in the previous commit.
As retrying a wait operation in WaitUntilObjectReadyWithHealthFunction
is not bound to returning an ErrorWithCodes anymore but rather to the
retriableError type, we can use that to indicate more explicitly, that
we want to retry waiting for a DNS object if it is in Error or Invalid
state.
This is now easier to understand as its cleaner and more explicit.

* Refactor `WaitUntilEtcdsReady` into `etcd.Wait`

This ties the wait func to the individual etcd component instances,
which is important for preventing unwanted consequences of stale
cache reads.

* Fix copyright year