Vendor golang/mock@v1.6.0 (#4210)
* Vendor golang/mock@v1.6.0

* Fix wrong DoAndReturn usages

There were a lot of DoAndReturn usages in our tests that didn't correctly use the
client func signatures. With golang/mock#558 this is now validated and tests
fail, if the signature of the provided func literal is not the same as the one of
the real func. This commit fixes our wrong usages to comply with the newest
gomock behavior.