Set `containerd` as default container runtime for k8s>=1.22
An empty `cri.name` field for a worker pool means all nodes are configured with `docker` as the container runtime. When users upgrade their shoots to k8s>=1.22 (or create new ones) the default is changed to `containerd`. This also means that starting with k8s 1.22 the shoot object will always contain an explicit configuration of the container runtime.