Replace CreateOrUpdate with GetAndCreateOr*Patch
CreateOrUpdate is likely to cause conflicts when used with a cached client,
i.e. when reading stale data (outdated resourceVersion) from the cache
before sending the update. Replace usages of it with GetAndCreateOr*Patch
which has similar semantics, except doing a patch instead of update, which
doesn't involve optimistic locking, so it cannot run into conflicts.
Merge patches without optimistic locking are fine for most use cases, where
we only set the spec of an object, plus a few labels and annotations.
When operating on ownerReferences, use strategic merge patches to run
without optimistic locking, but still be safe for concurrent updates.