Refactored support for node local dns.
Removed feature flag and introduced a shoot specific annotation instead.
The dns related network policies in the garden namespace and in the shoot control planes now always allow the traffic to the node local dns endpoints.
Node local dns should no longer have a cascading effect requiring shoots to enable it if it is enabled on the seed.