Eliminate retry.RetryOnConflict in scheduler
Seems like the reason for retrying the spec.seedName update is gone meanwhile
(gcm no longer sets status to pending for unscheduled shoots).
So, simplify the handling, use optimistic locking and use the default
rate limiter for the workqueue instead of custom rate limiter with baseDelay
of 15s to make retries faster.