Increased connect timeout of reversed vpn envoy sidecar.
The establishment of the reversed vpn connection seem to take around 5s.
This in conjunction with the fact that the reversed vpn connection may
need to be re-established quite often depending on the load on the istio
ingress gateway can cause the time frame without connectivity between
seed and shoot to be quite longer than the 1s connect timeout previously
configured.
This change increases the connect timeout with a good buffer so that
also `kubectl exec` and potentially also web hooks should no longer
fail, but rather stall until the vpn connection has been re-established.