Maintenance controller now respects workers CRI configuration (#4438)
* Respect CRI name when maintaining WorkerImages

Co-authored-by: Marco Voelz <marco.voelz@sap.com>

* Refactor MaintainMachineImages to also update shoot

We want to test the end result rather than the implementation details: If a shoot's worker pools get updated with a new image or not. In order to do this, we moved all logic dealing with selecting a MachineImageVersion and updating the shoot in a new method. This allows to make a method like selectUpdatedMachineImages private again, as it is an implementation detail only.

Co-authored-by: Marco Voelz <marco.voelz@sap.com>

* Update worker pools MachineImages independently

Without this fix two worker pools which selected the same MachineImage would
always get the same updated MachineImageVersion even if they differed in their
cri configuration. This could lead to a scenario where a worker pool was updated
to an ImageVersion which does not support the selected CRI.

Co-authored-by: Marco Voelz <marco.voelz@sap.com>

* Maintenance controller respects containerruntime

Co-authored-by: Marco Voelz <marco.voelz@sap.com>

* Move event recording and logging below shoot update

If the Shoot update would fail with an error, we don't see any events of things which didn't happen.

* Treat `cri:nil` in the same way like `cri.name:docker`

Co-authored-by: Marco Voelz <marco.voelz@sap.com>